refactor(read-models): BookDetail reads + Tag writes off DbContext (PR6b-3)
Completes BookDetailViewModel's migration off DbContext — the VM now holds
only presentation state and routes everything through IDispatcher.

Reads -> query handlers (AsNoTracking + DTO, C5):
- GetBookDetail: lifts InitializeAsync's 6-Include AsSplitQuery read + the
  flat projection; the BookDetail/WorkDetail/EditionDetail/CopyDetail/
  TagDetail/SeriesInfo records move from the VM into Application/Books.
- GetTagSuggestions: lifts the tag-autocomplete prefix read. The
  "exclude already-assigned" filter stays in the VM (it owns CurrentTags).

Writes -> command handlers, via a new shared TagResolver:
- AddTagToBook / RemoveTagFromBook lift the two inline tag DbContext blocks.
- TagResolver is the single owner of tag-name normalisation (trim +
  lowercase), mirroring AuthorResolver/PublisherResolver.

Tag dedup half of TD-15 closed: all three tag find-or-create sites now route
through TagResolver — AddTagToBook, MarkWishlistItemBought, and
BulkAddViewModel.EnsureFollowUpTagAsync. TD-15 stays Open for the unaddressed
concurrency race (deferred across all three resolvers) and the Publisher/Series
inline creates in the not-yet-migrated Add-flow VMs.

Also relocates BookFormatExtensions.DisplayName Web -> Application/Formatting
(4th formatter there) so the read handler can project it; _Imports already
carries the namespace so the Razor consumers need no change.

Tests: TagResolverTests cover the new shared seam directly; the relocated
read/tag-write handlers are exercised end-to-end by the existing
BookDetailViewModelTests (VM -> dispatcher -> handler) and
WishlistCommandHandlersTests. Full Bookcase suite green (681 passed, 1
skipped); clean --no-incremental build, no new warnings.
