+++ BookTracker.Application/Formatting/BookFormatExtensions.cs
@@ -1,7 +1,10 @@
 using BookTracker.Data.Models;
 
-namespace BookTracker.Web;
+namespace BookTracker.Application.Formatting;
 
+// Relocated from BookTracker.Web in PR6b-3 so the read-model handlers (which
+// project BookFormat into its display string) can reach it. The 4th formatter
+// to land in Application/Formatting after PR6a's two and PR6b-2's PartialDate.
 public static class BookFormatExtensions
 {
     public static string DisplayName(this BookFormat format) => format switch
