+++ BookTracker.Web/Components/Pages/Books/Detail.razor
@@ -1,4 +1,5 @@
 @page "/books/{BookId:int}"
+@using BookTracker.Application.Books
 @implements IAsyncDisposable
 @inject BookDetailViewModel VM
 @inject NavigationManager Nav
@@ -566,7 +567,7 @@
         if (!expandedWorkIds.Add(id)) expandedWorkIds.Remove(id);
     }
 
-    private List<BookDetailViewModel.WorkDetail> FilteredWorks(IReadOnlyList<BookDetailViewModel.WorkDetail> works)
+    private List<WorkDetail> FilteredWorks(IReadOnlyList<WorkDetail> works)
     {
         if (string.IsNullOrWhiteSpace(workFilter)) return works.ToList();
         var f = workFilter.Trim();
